Share Ember.EachView between HTMLBars and Handlebars.
Also fixup exports from ember-handlebars.
